+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/ContentBlockingController.java
@@ -102,6 +102,15 @@ public class ContentBlockingController {
     /** SocialTracking content has been loaded. */
     public static final int LOADED_SOCIALTRACKING_CONTENT = 0x00020000;
 
+    /** Email content has been blocked from loading. */
+    public static final int BLOCKED_EMAILTRACKING_CONTENT = 0x00400000;
+
+    /** EmailTracking content from the Disconnect level 1 has been loaded. */
+    public static final int LOADED_EMAILTRACKING_LEVEL_1_CONTENT = 0x00800000;
+
+    /** EmailTracking content from the Disconnect level 2 has been loaded. */
+    public static final int LOADED_EMAILTRACKING_LEVEL_2_CONTENT = 0x00000100;
+
     /**
      * Indicates that content that would have been blocked has instead been replaced with a shim.
      */
@@ -129,7 +138,9 @@ public class ContentBlockingController {
         Event.COOKIES_BLOCKED_TRACKER, Event.COOKIES_BLOCKED_SOCIALTRACKER,
         Event.COOKIES_BLOCKED_ALL, Event.COOKIES_PARTITIONED_FOREIGN,
         Event.COOKIES_BLOCKED_FOREIGN, Event.BLOCKED_SOCIALTRACKING_CONTENT,
-        Event.LOADED_SOCIALTRACKING_CONTENT, Event.REPLACED_TRACKING_CONTENT
+        Event.LOADED_SOCIALTRACKING_CONTENT, Event.REPLACED_TRACKING_CONTENT,
+        Event.LOADED_EMAILTRACKING_LEVEL_1_CONTENT, Event.LOADED_EMAILTRACKING_LEVEL_2_CONTENT,
+        Event.BLOCKED_EMAILTRACKING_CONTENT
       })
       public @interface LogEvent {}